backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
azure-ai-inference==1.0.0b3
requests==2.32.3
